*: re-organize file structure
